add stateful compose

+++ torchlight/transforms/noise.py
@@ -9,7 +9,7 @@ class GaussianNoise(object):
         self.sigma = sigma
 
     def __call__(self, img):
-        img_L = img + np.random.normal(0, self.sigma, img.shape)
+        img_L = img + np.random.normal(0, self.sigma, img.shape).astype(img.dtype)
         return img_L
